requirements/base.txt
Django==2.1.4
djangorestframework==3.9.0
djangorestframework_simplejwt==3.3
django-cors-headers==2.4.0
psycopg2==2.7.6.1
python-decouple==3.1
celery==4.2.1
redis==3.2.0
django-celery-results==1.0.4
requests==2.21.0
